+++ test/lib/remote_control/data/multiplexed_tv_connection_state_service_test.dart
@@ -136,4 +136,10 @@ class _RecordingCommandService implements RemoteCommandService {
   @override
   Future<TvDeviceInfo?> queryDeviceInfo({required TvDevice device}) async =>
       null;
+
+  @override
+  Future<void> pauseMonitoring({required TvDevice device}) async {}
+
+  @override
+  Future<void> resumeMonitoring({required TvDevice device}) async {}
 }